+++ src/main/webapp/WEB-INF/views/thymeleaf/productoEspecifico.html
@@ -75,9 +75,10 @@
                             Envios a todo el pais
                         </li>
                     </ul>
-                    <form method="POST" th:action="@{'/productoEspecifico/' + ${componenteEspecificoDto.id}}">
                         <button th:onclick="|agregarAlCarrito(${componenteEspecificoDto.id}, 1)|" class="btn btn-danger mt-3">Sumar al carrito</button>
-                    </form>
+                    <div th:id="'mensajeNotificacion-' + ${componenteEspecificoDto.id}" class="mensaje-notificacion"
+                         style="display:none;">
+                    </div>
                 </div>
             </div>
         </div>
@@ -118,6 +119,7 @@
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.6/dist/js/bootstrap.bundle.min.js" integrity="sha384-j1CDi7MgGQ12Z7Qab0qlWQ/Qqz24Gc6BM0thvEMVjHnfYGF0rmFCozFSxQBxwHKO" crossorigin="anonymous"></script>
         <script type="text/javascript" th:src="@{webjars/bootstrap/5.2.0/js/bootstrap.min.js}"></script>
         <script type="text/javascript" th:src="@{/js/main.js}"></script>
+        <script type="text/javascript" th:src="@{/js/agregarAlCarrito.js}"></script>
         <script th:src="@{/webjars/bootstrap/5.2.0/js/bootstrap.bundle.min.js}"></script>
         <script src="https://cdn.jsdelivr.net/npm/swiper@10/swiper-bundle.min.js"></script>
         <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/js/bootstrap.bundle.min.js"></script>
